requirements.txt
streamlit
joblib
numpy
pillow
pandas
scikit-learn